First time rendering planet position data
Now can see all six planet positions for the first frame of the animation. Using the latest format for the planet data JSON from HD110067 repo, EXCEPT that I had to delete most of the frames to bring the file size down to be able to commit to the repo. It should be looping in the original so need a smarter way to just get data I need later. The package-lock file happened when I tried to set up npm. Don't totally get if I need it or not.