Hover to filter feature first pass
Now when you hover on a line between planets in UI, the orbit animation only shows the orbit and planets for those planets. If nothing is hovered, it shows all planets. There is a bit of delay when hover/unhover due to how redrawing the orbit when it happens. Planning to switch to pre-making the static layers for each possible pair and then swapping in/out those images instead of redrawing the orbit points to see if that helps with performance.