Orbit filter switch time improved
Now when you hover/unhover over a planet pair, the switch is immediate. This is because at the beginning of the app I am creating graphics of the orbits for each possible pair of orbits, and one more for all orbits being visible. During the app, I just need to post one of the pair images, instead of having to draw all the ellipses of each orbit each time.